+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 
 ## [Unreleased]
 
+### Fixed
+
+- Generate static collection menus prebuild script: skip collections that are not included according to the config.
+
 
 
 ## [1.5.0] – 2024-08-28